# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Topic Platform",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/packs": {
      "get": {
        "summary": "Packs",
        "operationId": "packs_api_packs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/packs/{pack_id}/example": {
      "get": {
        "summary": "Example",
        "operationId": "example_api_packs__pack_id__example_get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "summary": "Runs",
        "operationId": "runs_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Start Run",
        "operationId": "start_run_api_runs_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_-]{1,100}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Run",
        "operationId": "run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/decisions": {
      "post": {
        "summary": "Decide",
        "operationId": "decide_api_runs__run_id__decisions_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/metrics": {
      "get": {
        "summary": "Metrics",
        "operationId": "metrics_api_runs__run_id__metrics_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/report.md": {
      "get": {
        "summary": "Report",
        "operationId": "report_api_runs__run_id__report_md_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Index",
        "operationId": "index__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DecisionRequest": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "criterion_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
            "title": "Criterion Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "accept",
              "reject",
              "edit"
            ],
            "title": "Action"
          },
          "actor": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Actor"
          },
          "note": {
            "type": "string",
            "maxLength": 3000,
            "title": "Note",
            "default": ""
          },
          "edited_suggestion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3000
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Suggestion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "criterion_id",
          "action",
          "actor"
        ],
        "title": "DecisionRequest"
      },
      "Document": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "version": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Version"
          },
          "role": {
            "type": "string",
            "enum": [
              "basis",
              "target"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 30000,
            "minLength": 1,
            "title": "Content"
          },
          "complete": {
            "type": "boolean",
            "title": "Complete",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "version",
          "role",
          "content"
        ],
        "title": "Document"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RunRequest": {
        "properties": {
          "pack_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
            "title": "Pack Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "model"
            ],
            "title": "Mode",
            "default": "demo"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "default": "새 검토"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 2,
            "title": "Documents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pack_id",
          "documents"
        ],
        "title": "RunRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}